refactor SuggestionContext and SuggestionService to improve package structure and evaluation logic
