+++ src/main/java/depromeet/lessonfour/server/report/domain/service/SuggestionService.java
@@ -1,30 +1,41 @@
 package depromeet.lessonfour.server.report.domain.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import depromeet.lessonfour.server.report.domain.policy.SuggestionPolicy;
-import depromeet.lessonfour.server.report.domain.vo.SuggestionType;
-import depromeet.lessonfour.server.report.domain.vo.monthly.MonthlyActivityReport;
-import depromeet.lessonfour.server.report.domain.vo.monthly.MonthlyToiletReport;
-import depromeet.lessonfour.server.report.domain.vo.weekly.WeeklyActivityReport;
-import depromeet.lessonfour.server.report.domain.vo.weekly.WeeklyToiletReport;
+import depromeet.lessonfour.server.report.domain.vo.activity.MonthlyActivityReport;
+import depromeet.lessonfour.server.report.domain.vo.activity.WeeklyActivityReport;
+import depromeet.lessonfour.server.report.domain.vo.suggestion.SuggestionContext;
+import depromeet.lessonfour.server.report.domain.vo.suggestion.SuggestionRule;
+import depromeet.lessonfour.server.report.domain.vo.suggestion.SuggestionType;
+import depromeet.lessonfour.server.report.domain.vo.toilet.MonthlyToiletReport;
+import depromeet.lessonfour.server.report.domain.vo.toilet.WeeklyToiletReport;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class SuggestionService {
 
-  private final SuggestionPolicy suggestionPolicy;
+  private final List<SuggestionRule> rules;
 
   public List<SuggestionType> suggest(
       WeeklyActivityReport weeklyActivityReport, WeeklyToiletReport weeklyToiletReport) {
-    return suggestionPolicy.evaluateWeekly(weeklyActivityReport, weeklyToiletReport);
+    SuggestionContext context = SuggestionContext.weekly(weeklyActivityReport, weeklyToiletReport);
+    List<SuggestionType> result = new ArrayList<>();
+    rules.forEach(rule -> rule.evaluate(context, result));
+
+    return result;
   }
 
   public List<SuggestionType> suggest(
       MonthlyActivityReport monthlyActivityReport, MonthlyToiletReport monthlyToiletReport) {
-    return suggestionPolicy.evaluateMonthly(monthlyActivityReport, monthlyToiletReport);
+    SuggestionContext context =
+        SuggestionContext.monthly(monthlyActivityReport, monthlyToiletReport);
+    List<SuggestionType> result = new ArrayList<>();
+    rules.forEach(rule -> rule.evaluate(context, result));
+
+    return result;
   }
 }
