+++ src/main/java/depromeet/lessonfour/server/report/domain/vo/suggestion/SuggestionContext.java
@@ -1,12 +1,16 @@
-package depromeet.lessonfour.server.report.domain.vo;
+package depromeet.lessonfour.server.report.domain.vo.suggestion;
 
 import java.util.List;
 
-import depromeet.lessonfour.server.report.domain.vo.daily.DailyActivityReport;
-import depromeet.lessonfour.server.report.domain.vo.monthly.MonthlyActivityReport;
-import depromeet.lessonfour.server.report.domain.vo.monthly.MonthlyToiletReport;
-import depromeet.lessonfour.server.report.domain.vo.weekly.WeeklyActivityReport;
-import depromeet.lessonfour.server.report.domain.vo.weekly.WeeklyToiletReport;
+import depromeet.lessonfour.server.report.domain.vo.activity.DailyActivityReport;
+import depromeet.lessonfour.server.report.domain.vo.activity.FoodEvaluation;
+import depromeet.lessonfour.server.report.domain.vo.activity.MonthlyActivityReport;
+import depromeet.lessonfour.server.report.domain.vo.activity.StressEvaluation;
+import depromeet.lessonfour.server.report.domain.vo.activity.WaterEvaluation;
+import depromeet.lessonfour.server.report.domain.vo.activity.WeeklyActivityReport;
+import depromeet.lessonfour.server.report.domain.vo.toilet.MonthlyToiletReport;
+import depromeet.lessonfour.server.report.domain.vo.toilet.ToiletEvaluation;
+import depromeet.lessonfour.server.report.domain.vo.toilet.WeeklyToiletReport;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -53,19 +57,13 @@ public class SuggestionContext {
 
   public static SuggestionContext monthly(
       MonthlyActivityReport activityReport, MonthlyToiletReport toiletReport) {
-    List<DailyActivityReport> allReports = activityReport.getAllReports();
+    List<FoodEvaluation> foodEvaluations = activityReport.getAllFoodEvaluations();
 
-    List<FoodEvaluation> foodEvaluations =
-        allReports.stream().flatMap(daily -> daily.getFoodEvaluations().stream()).toList();
+    List<WaterEvaluation> waterEvaluations = activityReport.getAllWaterEvaluations();
 
-    List<WaterEvaluation> waterEvaluations =
-        allReports.stream().flatMap(daily -> daily.getWaterEvaluations().stream()).toList();
+    List<StressEvaluation> stressEvaluations = activityReport.getAllStressEvaluations();
 
-    List<StressEvaluation> stressEvaluations =
-        allReports.stream().map(DailyActivityReport::getStressEvaluation).toList();
-
-    List<ToiletEvaluation> toiletEvaluations =
-        toiletReport.dailyReports().stream().flatMap(daily -> daily.getItems().stream()).toList();
+    List<ToiletEvaluation> toiletEvaluations = toiletReport.getAllEvaluations();
 
     return SuggestionContext.builder()
         .foodEvaluations(foodEvaluations)
